refactor: extract modules registration to dedicated class
