+++ Chapter-2-modules-separation/Src/Fitnet/Program.cs
@@ -1,12 +1,7 @@
-using EvolutionaryArchitecture.Fitnet;
 using EvolutionaryArchitecture.Fitnet.Common.Api.ErrorHandling;
 using EvolutionaryArchitecture.Fitnet.Common.Core.SystemClock;
 using EvolutionaryArchitecture.Fitnet.Common.Infrastructure;
-using EvolutionaryArchitecture.Fitnet.Common.Infrastructure.Modules;
-using EvolutionaryArchitecture.Fitnet.Contracts.Api;
-using EvolutionaryArchitecture.Fitnet.Offers.Api;
-using EvolutionaryArchitecture.Fitnet.Passes.Api;
-using EvolutionaryArchitecture.Fitnet.Reports;
+using EvolutionaryArchitecture.Fitnet.Modules;
 
 var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddControllers();
@@ -16,8 +11,7 @@ builder.Services.AddSwaggerGen();
 builder.Services.AddSystemClock();
 builder.Services.AddExceptionHandling();
 builder.Services.AddCommonInfrastructure();
-
-RegisterModules(builder);
+builder.Services.AddModules(builder.Configuration);
 
 var app = builder.Build();
 
@@ -35,21 +29,10 @@ app.UseErrorHandling();
 
 app.MapControllers();
 
-app.RegisterContracts(Module.Contracts);
-app.RegisterPasses(Module.Passes);
-app.RegisterOffers(Module.Offers);
-app.RegisterReports(Module.Reports);
+app.RegisterModules();
 
 app.Run();
 
-static void RegisterModules(WebApplicationBuilder webApplicationBuilder)
-{
-    using var moduleAvailabilityChecker = ModuleAvailabilityChecker.Build(webApplicationBuilder.Configuration);
-    webApplicationBuilder.Services.AddContracts(Module.Contracts, webApplicationBuilder.Configuration, moduleAvailabilityChecker);
-    webApplicationBuilder.Services.AddPasses(Module.Passes, webApplicationBuilder.Configuration, moduleAvailabilityChecker);
-    webApplicationBuilder.Services.AddOffers(Module.Offers, webApplicationBuilder.Configuration, moduleAvailabilityChecker);
-    webApplicationBuilder.Services.AddReports(Module.Reports, moduleAvailabilityChecker);
-}
 
 namespace EvolutionaryArchitecture.Fitnet
 {
